[Bronze II] Title: 이상한 곱셈, Time: 4816 ms, Memory: 305676 KB -BaekjoonHub

+++ Java/백준/Bronze/1225. 이상한 곱셈/이상한 곱셈.java
@@ -18,7 +18,7 @@ public class Main {
             for (String b : B.split("")) {
                 int j = Integer.parseInt(b);
 
-                result += i*j;
+                result += (long) i*j;
             }
         }
